Exoeriment-1/requirements.txt
torch>=2.1
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
faiss-cpu>=1.7.4
pytorch-metric-learning>=2.3
mlflow>=2.10
tqdm>=4.65
